refactor(structure): split pure parsers into obsidian-markdown/ and MCP wiring into mcp-core/
Make the parser/IO layering physical with zero behavior change. Pure
git mv + import-specifier rewrites:

- Pure Obsidian/Markdown parsers -> src/vault-mcp/obsidian-markdown/
  (frontmatter, callouts [was callout-parser], headings [was
  heading-parser], links). No fs/db/MCP deps.
- MCP protocol surface -> src/vault-mcp/mcp-core/
  (mcp-router, tool-definitions, prompt-definitions).
- Tests move with their sources (co-located __tests__/).
- Top-level rule: folders are domains; only the entry point (server.ts)
  and its config.ts stay loose.

Also updates the AGENTS.md structure tree and two stale comment
cross-references to the renamed files. No config changes needed (all
build/test tooling uses src/** globs). 898 tests pass, build + lint green.
