+++ AGENTS.md
@@ -70,16 +70,21 @@ src/
   vault-mcp/
     server.ts                          # Entry point — config, mount routes, listen
     config.ts                          # Env-var loader + ServerConfig type (loadConfig)
-    mcp-router.ts                      # /mcp session routes + transport lifecycle
-    tool-definitions.ts                # MCP tool registrations + Zod schemas
-    prompt-definitions.ts              # MCP prompt registrations + Zod arg schemas
-    vault-operations/                  # Vault content read/write/patch
+    obsidian-markdown/                 # Pure Obsidian/Markdown parsers + transforms (no I/O)
+      frontmatter.ts                   # gray-matter parse/stringify + frontmatter merge
+      callouts.ts                      # Leading-callout parser (> [!type] blocks)
+      headings.ts                      # Shared H1–H6 section-span parser (read + patch)
+      links.ts                         # Link grammar: parse, extract, resolve (wikilinks + md)
+    vault-operations/                  # Vault content read/write/patch (filesystem I/O)
       vault-filesystem.ts              # Read/write/list/delete .md files; outline + section reads
       vault-patcher.ts                 # Surgical edits: heading-targeted patch + find-and-replace
       note-mover.ts                    # Move/rename a note + rewrite every vault-wide link to it
-      heading-parser.ts                # Shared H1–H6 section-span parser (read + patch)
       memory-store.ts                  # About Me/ heading-aware read/append/delete
       daily-notes.ts                   # Daily note config reader + path resolver
+    mcp-core/                          # MCP protocol surface
+      mcp-router.ts                    # /mcp session routes + transport lifecycle
+      tool-definitions.ts              # MCP tool registrations + Zod schemas
+      prompt-definitions.ts            # MCP prompt registrations + Zod arg schemas
     search/                            # SQLite FTS5 indexing + file watching
       search-index.ts                  # SQLite FTS5 factory (tags, folders, etc)
       file-watcher.ts                  # chokidar -> keeps index current
