+++ src/vault-mcp/mcp-core/__tests__/tool-definitions.test.ts
@@ -1,9 +1,9 @@
 import { describe, it, expect, beforeEach, vi } from "vitest"
 import { registerTools, TOOL_NAMES } from "../tool-definitions.js"
-import { loadConfig } from "../config.js"
+import { loadConfig } from "../../config.js"
 import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js"
-import type { SearchIndex } from "../search/search-index.js"
-import { logger } from "../../logger.js"
+import type { SearchIndex } from "../../search/search-index.js"
+import { logger } from "../../../logger.js"
 
 const ALL_TOOL_NAMES = Object.values(TOOL_NAMES)
