+++ src/vault-mcp/obsidian-markdown/frontmatter.ts
