+++ src/vault-mcp/obsidian-markdown/headings.ts
