+++ src/vault-mcp/obsidian-markdown/links.ts
