+++ src/vault-mcp/server.ts
@@ -9,7 +9,7 @@ import { createMemoryStore } from "./vault-operations/memory-store.js"
 import { startFileWatcher } from "./search/file-watcher.js"
 import { createOAuthProvider } from "./auth/oauth-provider.js"
 import { createOAuthRoutes } from "./auth/oauth-routes.js"
-import { createMcpRouter } from "./mcp-router.js"
+import { createMcpRouter } from "./mcp-core/mcp-router.js"
 import { loadConfig } from "./config.js"
 import { logger } from "../logger.js"
 import env from "env-var"
