+++ src/vault-mcp/vault-operations/__tests__/memory-store.test.ts
@@ -10,7 +10,7 @@ import {
 import { mkdtemp, rm, writeFile, mkdir, readFile } from "node:fs/promises"
 import { join } from "node:path"
 import { tmpdir } from "node:os"
-import { parseNote } from "../frontmatter.js"
+import { parseNote } from "../../obsidian-markdown/frontmatter.js"
 import { createMemoryStore } from "../memory-store.js"
 import { logger } from "../../../logger.js"
